refactor: move EntryList logic to its own component
